Refactor LXDConnection http requests.
All requests were basically doing the same things, so it made sense
to unify those things into a single method, and create a better
experience around working with the response from that request.

The unit tests mock out most of this interaction, so it's hard to
verify, via unit tests, that this working properly, but I do have a
small and sloppy set of integration tests (that need to be...integrated)
that seem to indicate all is well (I was able to create, read, and
delete containers).