Improve documentation with a new container example
I understand that the config specification for `create()` is beyond the
scope of the documentation, but I had to look directly in lxd's code to
figure how to create a container from a local image, something which, I
think, is a rather common usage of (py)lxd.

In this commit, I add an example of such config.

This additional example should be benefical to many other users.