Fix the ws4py connection handling for container execution
This branch does a number of things. First, it simplifies _StdinWebsocket,
as we don't need most of what existed there. It closes immediately
after connection (and the command doesn't execute unless stdin is
also connected).

_CommandWebsocketClient learned how to pre-emptively close a connection.
We were previously waiting for the command to time out, which is why
the execute call was taking so long. This fixes #182.

The while loop that waits for the commands to complete had too long of
a delay, and had too much logic for whether the loop should continue
to wait. That while loop was also simplified.