Allow multipart form uploads for images.
In some cases, a user may want to upload a manifest with a raw image,
and pylxd previously didn't support that. This patch adds the needed
functionality.

As I did this, I found a bug in Image.create where fingerprint isn't
always reliably determined, so I have deprecated the ability to pass
a wait parameter to Image.create.

Fixes issue #172