Make models resilient to new attributes in LXD
It's been a few times now (ref #188, #230) where we're in a tricky situation
because a new attribute in LXD makes pylxd crash. It's not fun because
it makes us *have* to update pylxd and LXD at the same time.

This commit makes models ignore `AttributeError` during `sync()`, which
should normally fix this problem once and for all.

Depends on #233