Fix Operation class to allow unknown attributes
lxd 3.0.0-beta{x} grew a new feature in that the operation metadata
includes a 'description' tag.  This breaks pylxd, and so this fix adds
the 'description' key, and also then makes the class more robust to
handle unknown attributes in the future.  Note that ALL the other
classes got this fix in 2.2.4.

Fixes bug: #284

Signed-off-by: Alex Kavanagh <alex@ajkavanagh.co.uk>