Enable decoding control over stdout/stderr on execute
This patch adds support to control the decoding of the stdout
and stderr when executing commands in a container.  It is possible to
force a particular decode, or just get the raw buffers.  The existing
behaviour of trying to decode to utf-8 is retained if no decode or
encoding paramters are passed to the execute method.

Tests are in the integration/test_containers.py file as it's tricky to
test properly in a unit test.

Signed-off-by: Alex Kavanagh <alex@ajkavanagh.co.uk>