Remove overzealous '_' -> '-' subsitution in _APINode
The fix for storage pools in the __getattr__() in _APINode has
affected the names of containers and snapshots, if '_' are used in
the name.  This reverts most of that fix, adds tests, and special cases
'storage_pools' -> 'storage-pools' as an attribute ONLY, and not as
a name of a item.

Fixes: #295
Signed-off-by: Alex Kavanagh <alex@ajkavanagh.co.uk>