Implementation of storage pools for pylxd
This fully completes the 'storage' and 'resources' api_extensions of lxd
and also introduces a patch method into the Client class and extends the
put and patch methods to the storage volume (as they'll be needed to do
sane modifications).

This commit doesn't have the documentation changes, nor the functional
testing changes -- it does have the unit tests.

Signed-off-by: Alex Kavanagh <alex@ajkavanagh.co.uk>