Add integration test runner for bionic 18.04
lxd 3.0+ needs a slightly different test runner that creates a client
key directly so that the verify test path can be executed.  It also
requires a default storage pool to be created for the root dirs for the
containers.  There will be two further commits to rename the test runner
files so that the bionic is the default, with xenial as an additional
test.

Signed-off-by: Alex Kavanagh <alex@ajkavanagh.co.uk>