Add integration tests that can run on fresh vm with LXD
This adds the run-integration-tests-in-a-vm which will minimally
configure a fresh VM (or bare-metal) with LXD installed, and then run
the pylxd integration tests.

Also simplify the existing integration tests to a single
'run-integration-tests-in-lxd' which is for running the integration
tests in an existing, already configured, machine with LXD, where the
tests are run lxd-in-lxd.

Signed-off-by: Alex Kavanagh <alex@ajkavanagh.co.uk>