Add PYLXD_WARNINGS env variable to be able to supress warnings
If the LXD server that pylxd is connected to supports attributes on
objects that pylxd doesn't yet support, then a warning is issued using
the `warnings` module.  This can fill logs with annoying warnings.  So
this patch adds the ability to set an env variable PYLXD_WARNINGS to
'none' to suppress all the warnings, or to 'always' to get the existing
behaviour.  The new behavior is to issue a warning once for each
instance of an attribute that isn't known for each object.

Closes: #301
Signed-off-by: Alex Kavanagh <alex.kavanagh@canonical.com>