Updated example of profile updating and saving
It was discovered that using profile.config and profile.devices as left-hand-sides to a .update function was setting them to "none" and the profile.update() was not actually saving.  profile.save() was also causing the profile settings to be blanked out due to the NoneTypes being set to profile.config and profile.devices.  Also expanded profile attribute descriptions to explicitly note types