test_networks: Avoid triggering volatile.bridge.hwaddr
LXD now generates a stable MAC address for networks.
This is placed in volatile.bridge.hwaddr unless the network has a
pre-configured address.

This makes the current network test fail as it assumes that the config
returned after creation will be identical to that provided during
creation.

There are two ways of fixing this:

 - Strip all volatile keys
 - Provide a pre-set bridge.hwaddr key

This implements the latter.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>