pylxd/models/instance: only pass recursion if not 0
This fixes `tox -e coverage`. This was not caught when introduced by
PR #551 because we our CI job was still only triggering on the old
branch name. The CI job was fixed afterward by commit 094bab4090776468.

The decision to omit the recursion param was taken because that's easier
than trying to fix the mocked HTTP responses.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>